+++ backend/src/main/java/codesquad/airbnb/accomodation/repository/AccommodationRepositoryCustom.java
@@ -1,4 +1,4 @@
-package codesquad.airbnb.accomodation;
+package codesquad.airbnb.accomodation.repository;
 
 import codesquad.airbnb.accomodation.dto.PriceAndCount;
