+++ backend/src/main/java/codesquad/airbnb/accomodation/repository/AccommodationRepositoryCustomImpl.java
@@ -1,4 +1,4 @@
-package codesquad.airbnb.accomodation;
+package codesquad.airbnb.accomodation.repository;
 
 import codesquad.airbnb.accomodation.domain.QAccommodation;
 import codesquad.airbnb.accomodation.dto.PriceAndCount;
